refactor(#31): SelectMain 컴포넌트를 사용하는 개발자가 몰라도 되는 맥락에 대한 선언적 분리
